+++ mapie/conformity_scores/interface.py
@@ -1,10 +1,11 @@
 from abc import ABCMeta, abstractmethod
 from typing import Optional
 
-import numpy as np
 from sklearn.base import BaseEstimator
 
 from numpy.typing import NDArray
+
+from mapie.utils import _compute_quantile
 
 
 class BaseConformityScore(metaclass=ABCMeta):
@@ -126,33 +127,13 @@ class BaseConformityScore(metaclass=ABCMeta):
         NDArray of shape (1, n_alpha) or (n_samples, n_alpha)
             The quantiles of the conformity scores.
         """
-        n_ref = conformity_scores.shape[1 - axis]
-        n_calib: int = np.min(np.sum(~np.isnan(conformity_scores), axis=axis))
-        signed = 1 - 2 * reversed
-
-        # Adapt alpha w.r.t upper/lower : alpha vs. 1-alpha
-        alpha_ref = (1 - 2 * alpha_np) * reversed + alpha_np
-
-        # Adjust alpha w.r.t quantile correction
-        alpha_cor = np.ceil(alpha_ref * (n_calib + 1)) / n_calib
-        alpha_cor = np.clip(alpha_cor, a_min=0, a_max=1)
-
-        # Compute the target quantiles:
-        # If unbounded is True and alpha is greater than or equal to 1,
-        # the quantile is set to infinity.
-        # Otherwise, the quantile is calculated as the corrected lower quantile
-        # of the signed conformity scores.
-        quantile = signed * np.column_stack(
-            [
-                np.nanquantile(
-                    signed * conformity_scores, _alpha_cor, axis=axis, method="lower"
-                )
-                if not (unbounded and _alpha >= 1)
-                else np.inf * np.ones(n_ref)
-                for _alpha, _alpha_cor in zip(alpha_ref, alpha_cor)
-            ]
+        return _compute_quantile(
+            conformity_scores,
+            alpha_np,
+            axis=axis,
+            reverse=reversed,
+            unbounded=unbounded,
         )
-        return quantile
 
     @abstractmethod
     def predict_set(self, X: NDArray, alpha_np: NDArray, **kwargs):